requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
coloredlogs==15.0.1
fastapi==0.119.1
flatbuffers==25.9.23
h11==0.16.0
humanfriendly==10.0
idna==3.11
numpy==2.3.4
onnx==1.19.1
onnxruntime==1.23.1
packaging==25.0
pillow==12.0.0
protobuf==6.33.0
pydantic==2.12.3
pydantic_core==2.41.4
python-multipart==0.0.20
setuptools==80.9.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
